+++ tad/city.h
@@ -2,11 +2,11 @@
 #define CITY_H
 
 typedef struct {
-    int x;
-    int y; // coordonnées GPS
+    double x;
+    double y; // coordonnées GPS
     int id;
 }City ;
 
-City* createCity(int id, int xCoor, int yCoor);
+City* createCity(int id, double xCoor, double yCoor);
 void freeCity(City* city);
 #endif //CITY_H
