+++ tad/matrix.h
@@ -5,19 +5,19 @@
 
 typedef struct {
     int dimension;
-    int **matrix;
+    double **matrix;
 } Matrix;
 
 typedef struct {
     int dimension;
-    int bestDistance;
+    double bestDistance;
     int *bestPath;
 } Results;
 
 
 Matrix* MatrixCreate(int dimension);
 void freeMatrix(Matrix* m);
-void setDistance(Matrix* m, int i, int j, int distance);
-int getDistance(Matrix* m, int i, int j);
+void setDistance(Matrix* m, int i, int j, double distance);
+double getDistance(Matrix* m, int i, int j);
 
 #endif //MATRIX_H
